// projects/dashboard/src/app/interfaces/oanda/oanda.ts
export type TakeProfitOrder = Omit<
  Order,
  | 'instrument'
  | 'units'
  | 'priceBound'
  | 'positionFill'
  | 'takeProfitOnFill'
  | 'stopLossOnFill'
  | 'guaranteedStopLossOnFill'
  | 'trailingStopLossOnFill'
  | 'tradeClientExtensions'
  | 'distance'
  | 'trailingStopValue'
>;

/**
 * A StopLossOrder is an order that is linked to an open Trade and created with a
 * price threshold. The Order will be filled (closing the Trade) by the first price
 * that is equal to or worse than the threshold. A StopLossOrder cannot be used to
 * open a new Position.
 */
export type StopLossOrder = Omit<
  Order,
  | 'instrument'
  | 'units'
  | 'priceBound'
  | 'positionFill'
  | 'takeProfitOnFill'
  | 'stopLossOnFill'
  | 'guaranteedStopLossOnFill'
  | 'trailingStopLossOnFill'
  | 'tradeClientExtensions'
  | 'tradeState'
  | 'initialMarketPrice'
  | 'trailingStopValue'
>;
// ...
export type TrailingStopLossOrder = Omit<
  Order,
  | 'instrument'
  | 'units'
  | 'price'
  | 'priceBound'
  | 'positionFill'
  | 'takeProfitOnFill'
  | 'stopLossOnFill'
  | 'guaranteedStopLossOnFill'
  | 'trailingStopLossOnFill'
  | 'tradeClientExtensions'
  | 'tradeState'
  | 'initialMarketPrice'
>;

/** The current state of the Order. */
type OrderState = 'PENDING' | 'FILLED' | 'TRIGGERED' | 'CANCELLED';

interface ClientExtensions {
  /**
   * The Client ID of the Order/Trade
   */
  id: string;

  /**
   * A tag associated with the Order/Trade
   */
  tag: string;

  /**
   * A comment associated with the Order/Trade
   */
  comment: string;
}
// ...
export interface Order {
// ...
  instrument: string;
// ...
}
// ...
type TradeState = 'OPEN' | 'CLOSED' | 'CLOSE_WHEN_TRADEABLE';
// ...
export type TradeStateFilter =
  | 'OPEN'
  | 'CLOSED'
  | 'CLOSE_WHEN_TRADEABLE'
  | 'ALL';

/** Get a list of Orders for an Account */
export interface Trades {
  /** The list of Order detail objects */
  trades: Trade[];

  /** The ID of the most recent Transaction created for the Account. */
  lastTransactionID: string;
}

export interface SpecifiedTrade {
  /** The details of the requested trade */
  trade: Trade;

  /** The ID of the most recent Transaction created for the Account. */
  lastTransactionID: string;
}

/**
 * The specification of a Trade within an Account. This includes the
 * full representation of the Trade's dependent Orders in addition to
 * the IDs of those Orders.
 */
export interface Trade {
  /** The Trade's identifier, unique within the Trade's Account. */
  id: string;

  /** The Trade's Instrument. */
  instrument: string;

  /** The execution price of the Trade. */
  price: string;

  /** The date/time when the Trade was opened. */
  openTime: string;

  /** The current state of the Trade. */
  state: TradeState;

  /**
   * The initial size of the Trade. Negative values indicate a short Trade,
   * and positive values indicate a long Trade.
   */
  initialUnits: string;

  /**
   * The margin required at the time the Trade was created. Note, this is the
   * ‘pure' margin required, it is not the ‘effective' margin used that
   * factors in the trade risk if a GSLO is attached to the trade.
   */
  initialMarginRequired: string;

  /**
   * The number of units currently open for the Trade. This value is reduced
   * to 0.0 as the Trade is closed.
   */
  currentUnits: string;

  /** The total profit/loss realized on the closed portion of the Trade. */
  realizedPL: string;

  /** The unrealized profit/loss on the open portion of the Trade. */
  unrealizedPL: string;

  /** Margin currently used by the Trade. */
  marginUsed: string;

  /**
   * The average closing price of the Trade. Only present if the Trade has
   * been closed or reduced at least once.
   */
  averageClosePrice: string;

  /** The IDs of the Transactions that have closed portions of this Trade. */
  closingTransactionIDs?: string[];

  /** The financing paid/collected for this Trade. */
  financing: string;

  /** The dividend adjustment paid for this Trade. */
  dividendAdjustment: string;

  /**
   * The date/time when the Trade was fully closed. Only provided for Trades
   * whose state is CLOSED.
   */
  closeTime: string;

  /** The client extensions of the Trade. */
  clientExtensions: ClientExtensions;

  /**
   * Full representation of the Trade's Take Profit Order, only provided if
   * such an Order exists.
   */
  takeProfitOrder: TakeProfitOrder;

  /**
   * Full representation of the Trade's Stop Loss Order, only provided if such
   * an Order exists.
   */
  stopLossOrder: StopLossOrder;

  /**
   * Full representation of the Trade's Trailing Stop Loss Order, only
   * provided if such an Order exists.
   */
  trailingStopLossOrder: TrailingStopLossOrder;
}